+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 RECENT CHANGES
 --------------
 
+- New option to avoid passing PDFs already containing text to the OCR engine
 - New option to add a trivial value to the output filename (like a date)
 23 Aug 2015: v1.04
 - Fixed multiple problems with spaces in filenames and exclusion patterns